+++ packages/period-selector-dialog/src/modules/FixedPeriodsGenerator.js
