+++ packages/period-selector-dialog/src/modules/RelativePeriodsGenerator.js
