+++ packages/period-selector-dialog/src/modules/periodTypes.js
