+++ packages/period-selector-dialog/src/styles/PeriodFilter.style.js
